requirements.txt
python-telegram-bot==20.6
yt-dlp==2023.7.6
imageio-ffmpeg==0.6.0